[chore][confighttp] Put client and server code into separate files
